Deleted SogoLogin, reorganize
